- Update refresh token flow when app enter foreground
